src/requirements.txt
Alfred-Workflow>=1.37.2
requests>=2.22.0
